SystemMessage Update - tnx kombat
With the capability to add own npcs to the database we should also be able to see their names in system messages. With the current system, where the names are always added by npc ID, which can be only a template id for the npc, we are not.
With this patch it is detected if the npc template is custom, and if it is, then the npc name is added by raw string.
Same for skills.

Now new method added to SystemMessage: addCharName() which takes any L2Character.
Depending on the given reference the name will added by npc ID, summon ID or by PC visibleName.

This saves us from adding a instanceof-if to every skill handler and so on.
Additional, with this patch, less bandwith is consumed since in many places of code NPCs were still added by String instead of ID.